--- src/components/Items.js
@@ -1,6 +1,0 @@
-export default [
-  { id: 1, item: 'Home', link: '/home' },
-  { id: 2, item: 'About Us', link: '/about' },
-  { id: 3, item: 'Contact Us', link: '/contactus' },
-  { id: 4, item: 'Login', link: '/login' }
-];
