+++ src/components/contact/contact.css
@@ -31,10 +31,6 @@ a:hover {
   text-decoration: none;
 }
 
-
-}
-
-
 ul,
 li {
   margin: 0px;
