--- src/components/product.js
@@ -1,74 +1,0 @@
-export default [
-  {
-    amount: '#50.00',
-    id: '1',
-    name: 'Bell Pepper',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084141/product-1_my7jm8.jpg'
-  },
-  {
-    amount: '#45.00',
-    id: '2',
-    name: 'Straw Berry',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084143/product-2_bbaozz.jpg'
-  },
-  {
-    amount: '#55.00',
-    id: '3',
-    name: 'Green Pea',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084140/product-3_dohofc.jpg'
-  },
-  {
-    amount: '#65.00',
-    id: '4',
-    name: 'Purple Garbage',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084140/product-4_lpullr.jpg'
-  },
-  {
-    amount: '#90.00',
-    id: '5',
-    name: 'Purple Garbage',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084140/product-4_lpullr.jpg'
-  },
-  {
-    amount: '#50.00',
-    id: '6',
-    name: 'Bell Pepper',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084141/product-1_my7jm8.jpg'
-  },
-  {
-    amount: '#45.00',
-    id: '7',
-    name: 'Straw Berry',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084143/product-2_bbaozz.jpg'
-  },
-  {
-    amount: '#55.00',
-    id: '8',
-    name: 'Green Pea',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084140/product-3_dohofc.jpg'
-  },
-  {
-    amount: '#65.00',
-    id: '9',
-    name: 'Purple Garbage',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084140/product-4_lpullr.jpg'
-  },
-  {
-    amount: '#90.00',
-    id: '10',
-    name: 'Purple Garbage',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084140/product-4_lpullr.jpg'
-  },
-  {
-    amount: '#50.00',
-    id: '11',
-    name: 'Bell Pepper',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084141/product-1_my7jm8.jpg'
-  },
-  {
-    amount: '#90.00',
-    id: '12',
-    name: 'Purple Garbage',
-    src: 'https://res.cloudinary.com/pauliski/image/upload/v1590084140/product-4_lpullr.jpg'
-  }
-];
